Move SaveMode tests into their own suite.
